clean light test comments

+++ Game-build/Assets/shaders/pointlight.fs.glsl
@@ -1,7 +1,5 @@
-
-
 #version 300 es
-precision mediump float;
+precision highp float;
 
 
 in vec3 vPosition_vs;
